# backend/internal/http/docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenShop API
  version: 1.0.0
  description: |
    Horizontally scalable e-commerce API. All responses use an envelope:
    successful responses carry `data` (and `meta` for lists), errors carry
    `error: { code, message }`.

    Authentication is bearer JWT. Send `Idempotency-Key` on POST /orders and
    POST /payments to make retries safe.
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in with email or phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string }
                password: { type: string }
      responses:
        "200": { $ref: "#/components/responses/AuthResult" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate a refresh token
      description: |
        Refresh tokens are single-use and belong to a family. Reusing a rotated
        token is treated as theft and revokes the whole family.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200": { $ref: "#/components/responses/AuthResult" }
        "401": { $ref: "#/components/responses/Error" }
...
  /auth/password/change:
    post:
      tags: [Auth]
      summary: Change your password
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "200": { description: Changed }
        "401": { $ref: "#/components/responses/Error" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current tokens
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "204": { description: Revoked }
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Current profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/User" }
...
  /products:
    get:
      tags: [Catalog]
      summary: List or search products
      parameters:
        - { in: query, name: categoryId, schema: { type: string, format: uuid } }
        - { in: query, name: keyword, schema: { type: string } }
        - { in: query, name: sort, schema: { type: string, enum: [newest, price_asc, price_desc] } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: A page of products
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Product" }
                      meta: { $ref: "#/components/schemas/Meta" }
  /products/{id}:
    get:
      tags: [Catalog]
      summary: Get a product
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The product
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/Error" }
...
  /orders:
    post:
      tags: [Orders]
      summary: Checkout the cart
      description: Reserves stock atomically and creates a pending order.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: header, name: Idempotency-Key, schema: { type: string } }
      responses:
        "201":
          description: The order
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    get:
      tags: [Orders]
      summary: List orders (admins see all)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: status, schema: { $ref: "#/components/schemas/OrderStatus" } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 10 } }
      responses:
        "200":
          description: A page of orders
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Order" }
                      meta: { $ref: "#/components/schemas/Meta" }
  /orders/{id}/invoice:
    get:
      tags: [Orders]
      summary: Download the order invoice as a PDF
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: A PDF invoice
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "403": { description: Not the order owner }
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get an order
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The order
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/Error" }
...
  /ops/categories:
    post:
      tags: [Ops]
      summary: Create a category
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                slug: { type: string }
                parentId: { type: string, format: uuid }
                sort: { type: integer }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Category" }
  /ops/categories/{id}:
    patch:
      tags: [Ops]
      summary: Update a category
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                slug: { type: string }
                parentId: { type: string, format: uuid }
                sort: { type: integer }
      responses:
        "200":
          description: The updated category
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Category" }
  /ops/products:
    post:
      tags: [Ops]
      summary: Create a product
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Product" }
  /ops/products/{id}:
    get:
      tags: [Ops]
      summary: Get one product with its variants
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The product
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Product" }
    patch:
      tags: [Ops]
      summary: Update a product
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Product" }
  /ops/uploads:
    post:
      tags: [Ops]
      summary: Upload an image
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      key: { type: string }
                      url: { type: string, format: uri }
  /products/{id}/reviews:
    get:
      tags: [Catalog]
      summary: List reviews for a product
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 10 } }
      responses:
        "200":
          description: A page of reviews
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Review" }
                      meta: { $ref: "#/components/schemas/Meta" }
    post:
      tags: [Catalog]
      summary: Create a review (one per user per product)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                title: { type: string }
                body: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Review" }
        "409": { $ref: "#/components/responses/Error" }
  /reviews/{id}:
    patch:
      tags: [Catalog]
      summary: Update your review
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                title: { type: string }
                body: { type: string }
      responses:
        "200": { description: Updated }
    delete:
      tags: [Catalog]
      summary: Delete a review (owner or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
...
  /ops/coupons:
    get:
      tags: [Ops]
      summary: List coupons
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Coupons }
    post:
      tags: [Ops]
      summary: Create a coupon
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, discountType, discountValue]
              properties:
                code: { type: string }
                description: { type: string }
                discountType: { type: string, enum: [percent, fixed] }
                discountValue: { type: integer, format: int64 }
                minSubtotalCents: { type: integer, format: int64 }
                maxDiscountCents: { type: integer, format: int64 }
                usageLimit: { type: integer }
                perUserLimit: { type: integer }
                startsAt: { type: string, format: date-time }
                endsAt: { type: string, format: date-time }
                active: { type: boolean }
      responses:
        "201": { description: Created }
  /ops/orders/{id}/refund:
    post:
      tags: [Ops]
      summary: Refund part or all of a paid order
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
                amountCents: { type: integer, format: int64, description: "Blank/0 refunds the remaining balance" }
                restock: { type: boolean, description: "Return inventory when the order becomes fully refunded" }
      responses:
        "200":
          description: The refunded order
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Order" }
        "409": { $ref: "#/components/responses/Error" }
  /orders/{id}/returns:
    get:
      tags: [Orders]
      summary: List the caller's return requests for an order
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Return requests }
    post:
      tags: [Orders]
      summary: Request a return for a delivered order
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "201": { description: The return request }
        "409": { $ref: "#/components/responses/Error" }
  /ops/returns:
    get:
      tags: [Ops]
      summary: List return requests
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Return requests }
  /ops/returns/{id}/approve:
    post:
      tags: [Ops]
      summary: Approve a return request
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: The approved return request }
  /ops/returns/{id}/reject:
    post:
      tags: [Ops]
      summary: Reject a return request
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: The rejected return request }
  /ops/products/{id}/variants:
    get:
      tags: [Ops]
      summary: List a product's variants
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Variants }
    post:
      tags: [Ops]
      summary: Create a variant (SKU)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                sku: { type: string }
                name: { type: string }
                priceCents: { type: integer, format: int64, description: "0 inherits the product price" }
                stock: { type: integer, minimum: 0 }
                attributes: { type: object, additionalProperties: { type: string } }
                sort: { type: integer }
                active: { type: boolean }
      responses:
        "201": { description: Created }
  /ops/variants/{id}:
    patch:
      tags: [Ops]
      summary: Update a variant
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sku: { type: string }
                name: { type: string }
                priceCents: { type: integer, format: int64 }
                stock: { type: integer }
                attributes: { type: object, additionalProperties: { type: string } }
                sort: { type: integer }
                active: { type: boolean }
      responses:
        "200": { description: Updated }
...
  /orders/{id}/complete:
    post:
      tags: [Orders]
      summary: Confirm receipt of a shipped order
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Completed }
        "409": { $ref: "#/components/responses/Error" }
  /ops/orders/{id}/ship:
    post:
      tags: [Ops]
      summary: Mark a paid order as shipped
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                trackingNo: { type: string }
      responses:
        "200": { description: Shipped }
        "409": { $ref: "#/components/responses/Error" }
  /ops/orders/{id}/complete:
    post:
      tags: [Ops]
      summary: Mark a shipped order as completed
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Completed }
  /ops/stats:
    get:
      tags: [Ops]
      summary: Merchant dashboard (revenue, order counts, recent orders)
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Dashboard }
...
  /ops/currencies/{code}:
    put:
      tags: [Ops]
      summary: Set an exchange rate (micro units, base -> code)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: code, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rateMicro]
              properties:
                rateMicro: { type: integer, format: int64 }
      responses:
        "200": { description: Saved }
  /ops/audit-logs:
    get:
      tags: [Ops]
      summary: Read the audit trail
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: action, schema: { type: string } }
        - { in: query, name: actorId, schema: { type: string, format: uuid } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
      responses:
        "200": { description: A page of audit entries }
  /ops/settings:
    get:
      tags: [Ops]
      summary: List runtime settings
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Every runtime setting with its current and default value
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        key: { type: string }
                        group: { type: string }
                        type: { type: string, enum: [string, int, bool] }
                        value: { type: string }
                        default: { type: string }
                        description: { type: string }
    put:
      tags: [Ops]
      summary: Update runtime settings (applied without a restart)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: { type: string }
      responses:
        "200": { description: The resolved settings }
        "400": { $ref: "#/components/responses/Error" }
...
  /ops/coupons/{id}:
    patch:
      tags: [Ops]
      summary: Update or deactivate a coupon
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description: { type: string }
                discountValue: { type: integer, format: int64 }
                minSubtotalCents: { type: integer, format: int64 }
                maxDiscountCents: { type: integer, format: int64 }
                usageLimit: { type: integer }
                perUserLimit: { type: integer }
                active: { type: boolean }
      responses:
        "200": { description: Updated }
  /ops/reviews:
    get:
      tags: [Ops]
      summary: List reviews across products for moderation
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: page, schema: { type: integer, default: 1 } }
      responses:
        "200": { description: A page of reviews }
  /ops/inventory/low-stock:
    get:
      tags: [Ops]
      summary: Products and variants at or below the low-stock threshold
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Low-stock items }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
    AuthResult:
      description: Authenticated session
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  accessToken: { type: string }
                  refreshToken: { type: string }
                  expiresIn: { type: integer }
...
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
        meta: { $ref: "#/components/schemas/Meta" }
    Meta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        phone: { type: string }
        name: { type: string }
        role: { type: string, enum: [customer, admin] }
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        parentId: { type: string }
        sort: { type: integer }
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        categoryId: { type: string }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        priceCents: { type: integer, format: int64 }
        currency: { type: string }
        coverImage: { type: string }
        images: { type: array, items: { type: string } }
        status: { type: string, enum: [draft, published, archived] }
        stock: { type: integer }
        weightGrams: { type: integer }
        variants:
          type: array
          items: { $ref: "#/components/schemas/Variant" }
        rating: { type: number }
        reviewCount: { type: integer }
...
    OrderStatus:
      type: string
      enum: [pending_payment, paid, cancelled, shipped, completed, refunded]
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderNo: { type: string }
        status: { $ref: "#/components/schemas/OrderStatus" }
        currency: { type: string }
        subtotalCents: { type: integer, format: int64 }
        discountCents: { type: integer, format: int64 }
        couponCode: { type: string }
        shippingCents: { type: integer, format: int64 }
        taxCents: { type: integer, format: int64 }
        shippingMethod: { type: string }
        totalCents: { type: integer, format: int64 }
        paymentId: { type: string }
        shippingAddress: { $ref: "#/components/schemas/Address" }
        trackingNo: { type: string }
        shippedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        paidAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              productId: { type: string, format: uuid }
              variantId: { type: string, format: uuid }
              variantName: { type: string }
              sku: { type: string }
              title: { type: string }
              priceCents: { type: integer, format: int64 }
              quantity: { type: integer }
              subtotal: { type: integer, format: int64 }
...